Fix comment issues
