+++ sdk/index.js
@@ -1,5 +1,5 @@
 const { createChatBox, setChatBoxDivId } = require('./chat/chat_handlers');
-const { goToLoginForm, initializeAuthenticationContainer } = require('./login_form/login_form_handler');
+const { goToLoginForm, initializeAuthenticationContainer } = require('./login_form/authentication_handler');
 const { configureInfiniteScroll } = require('./infinite_scroll/infinite_scroll');
 const { configureInfiniteScrollV2 } = require('./infinite_scroll/infinite_scroll_v2');
